refactor: move eox-core to users dir
