--- eox_nelp/eox_core/__init__.py
@@ -1,3 +1,0 @@
-"""
-EOX Core extension API package.
-"""
