--- eox_nelp/eox_core/api/__init__.py
@@ -1,3 +1,0 @@
-"""
-API package.
-"""
