--- eox_nelp/eox_core/api/v1/__init__.py
@@ -1,3 +1,0 @@
-"""
-API v1 package.
-"""
