--- eox_nelp/eox_core/tests/__init__.py
@@ -1,3 +1,0 @@
-"""
-Tests for the eox_core module.
-"""
