+++ eox_nelp/users/__init__.py
@@ -1,0 +1,3 @@
+"""
+Users app package.
+"""
