+++ eox_nelp/users/urls.py
@@ -1,9 +1,9 @@
 """
-Main URLs for eox-core.
+Main URLs for users.
 """
 from django.urls import include, path
 
 app_name = 'eox_nelp'  # pylint: disable=invalid-name
 urlpatterns = [
-    path('api/v1/', include('eox_nelp.eox_core.api.v1.urls', namespace='v1')),
+    path('v1/', include('eox_nelp.users.api.v1.urls', namespace='v1')),
 ]
